refactor(rust): eliminate comfy_table — all tables use custom renderer
Replace comfy_table with table_render::SimpleTable across entire codebase.
comfy_table wraps content instead of truncating, causing multi-line rows
on narrow terminals. The custom renderer guarantees single-line rows with
Unicode ellipsis truncation.

Converted 7 files:
- cmd_cleanup_costs.rs (cost history table)
- cmd_cleanup_costs2.rs (budget + recommendation tables)
- cmd_cleanup_ops.rs (orphaned resource table)
- cmd_env.rs (environment variable table)
- cmd_storage_ops.rs (storage account table)
- cmd_vm_ops.rs (VM detail property/value table)
- azlin-cli/table.rs (CLI table utilities)

Moved canonical SimpleTable implementation to azlin-cli crate,
re-exported from azlin crate. Removed comfy-table from all Cargo.toml.

2,536 tests pass, 0 failures. Zero comfy_table references remain.
